Fix image-url tag for the parallax image in SCSS
It's useful to read about how Rails handles assets in production.
The "fingerprinting" code that Rails adds to each file is to make sure
that when the file changes, Rails serves up the newest version of the
file. The fingerprint that gets added to the filename changes whenever
the file changes, so Rails can keep track of which version of the file
is the most recent just by storing the fingerprint every time it
compiles the files.

http://guides.rubyonrails.org/asset_pipeline.html#in-production